Pausing an animatable should pause all children
I've had need to pause the animations on an entire scene. If the model
is complex, manually attempting to do this will cause about a ~1ms
difference between different bones on the model.

Having this not recurse by default can also cause unexpected behavior
when calling `setPause` on a scene. The scene will stop rendering, but
unpausing will cause it to jump forward to the frame it would have been
on otherwise.